apps/crewai/requirements.txt
crewai>=0.22.0
langchain-google-genai>=1.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
